Packaging, profiling, layer improvments

+++ scripts/package.bat
@@ -1,11 +1,29 @@
-@echo ON
+@echo OFF
 
-call scripts\get_build_config.bat %1
+call scripts\get_build_config.bat --release
 
-set "package_dir=package\DrkCraft"
+set /p version=<VERSION.txt
 
-rmdir /S /Q %package_dir%
+set "package_name=DrkCraft-%version%-win"
+set "package_dir=packages\%package_name%"
+
+echo Building package: %package_name%
+
+call scripts\clean.bat
+call scripts\build.bat --release
+if NOT %errorlevel% == 0 (
+    echo CMake build failed
+    exit /b 1
+)
+
+if exist %package_dir% rmdir /S /Q %package_dir%
 mkdir %package_dir%
 
-xcopy %exe% %package_dir%
-xcopy data %package_dir%\data\ /E /y /EXCLUDE:scripts\package_exclusion_list.txt
+echo Copying executable
+xcopy /E %exe% %package_dir%
+
+echo Copying data files
+xcopy /E data %package_dir%\data\
+
+echo Copying config files
+xcopy /E config %package_dir%\config\
